Modularize HID simulation and visual verification logic
This commit implements Phase 2 and Phase 4 of the project roadmap by
extracting HID event simulation and image verification into dedicated
modules.

- Created `scripts/hid_simulator.py` to encapsulate HID event logic.
- Created `scripts/image_verifier.py` for screenshot capture and template matching.
- Refactored `scripts/hid_verify.py` to use these new modules, improving maintainability.
- Updated `ROADMAP.md` to reflect the progress.
- Verified functionality via `scripts/run_ci.sh` against the mock UI.
